voice-summary/requirements.txt
# Voice Summary Hook - Python Dependencies

# ===== Quick Install =====

# For simple mode (default):
#   No installation needed! Just works on macOS.

# For Gemini mode (recommended):
#   pip install google-generativeai

# For custom mode with specific providers:
#   OpenAI:    pip install openai
#   Anthropic: pip install anthropic
#   Ollama:    pip install requests
#   ElevenLabs: pip install requests

# ===== All providers (if you want everything) =====
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
